move add_nice_names to data_utils and make output_power precommit compliant
